Improves volcano risk assessment
Added a new earthquake risk index calculation to the advanced strain analysis, improving the accuracy of regional risk assessment by correctly handling data types in numpy array filtering within `utils/advanced_strain_utils.py` and updating  `pages/climate_volcanoes.py` to utilize this new function.

+++ utils/advanced_strain_utils.py
@@ -556,9 +556,10 @@ def calculate_earthquake_risk_index(strain_data, region_name, earthquake_history
         try:
             # Filter for data in this region
             lat, lon = region_info["lat"], region_info["lon"]
+            # Filter data for this region with proper numpy usage
             regional_strain = strain_data[
-                (np.abs(strain_data["latitude"] - lat) < 5) & 
-                (np.abs(strain_data["longitude"] - lon) < 5)
+                (np.abs(strain_data["latitude"].astype(float) - lat) < 5) & 
+                (np.abs(strain_data["longitude"].astype(float) - lon) < 5)
             ]
             
             if not regional_strain.empty:
